20150318
POLRT working.
It was just a question of approximations in the zeros of the polynomial
matrix in the 2nd example.
It works with zeros given by POLRT on DET of A.
Added also ERpy that introduces binding to ER.py in order to reduce the
number of parameters given to the functions.